+++ crates/forge-runner/src/running.rs
@@ -8,7 +8,7 @@ use crate::gas::calculate_used_gas;
 use crate::sierra_casm_runner::SierraCasmRunner;
 use crate::test_case_summary::{Single, TestCaseSummary};
 use crate::{RunnerConfig, RunnerParams, TestCaseRunnable, CACHE_DIR};
-use anyhow::{anyhow, bail, ensure, Result};
+use anyhow::{bail, ensure, Result};
 use blockifier::execution::common_hints::ExecutionMode;
 use blockifier::execution::entry_point::{
     CallEntryPoint, CallType, EntryPointExecutionContext, ExecutionResources,
@@ -37,22 +37,16 @@ use cheatnet::runtime_extensions::io_runtime_extension::IORuntimeExtension;
 use cheatnet::state::{BlockInfoReader, CheatnetBlockInfo, CheatnetState, ExtendedStateReader};
 use itertools::chain;
 use runtime::{ExtendedRuntime, StarknetRuntime};
-use starknet::core::types::BlockTag::Latest;
-use starknet::core::types::{BlockId, MaybePendingBlockWithTxHashes};
+use starknet::core::types::BlockId;
 use starknet::core::utils::get_selector_from_name;
-use starknet::providers::jsonrpc::HttpTransport;
-use starknet::providers::{JsonRpcClient, Provider};
-use starknet_api::block::BlockNumber;
 use starknet_api::core::PatriciaKey;
 use starknet_api::core::{ContractAddress, EntryPointSelector};
 use starknet_api::deprecated_contract_class::EntryPointType;
 use starknet_api::hash::StarkHash;
 use starknet_api::patricia_key;
 use starknet_api::transaction::Calldata;
-use tokio::runtime::Runtime;
 use tokio::sync::mpsc::Sender;
 use tokio::task::JoinHandle;
-use url::Url;
 
 /// Builds `hints_dict` required in `cairo_vm::types::program::Program` from instructions.
 fn build_hints_dict<'b>(
@@ -193,13 +187,8 @@ fn build_syscall_handler<'a>(
     )
 }
 
-pub(crate) struct ForkInfo {
-    pub(crate) latest_block_number: Option<BlockNumber>,
-}
-
 pub struct RunResultWithInfo {
     pub(crate) run_result: Result<RunResult, RunnerError>,
-    pub(crate) fork_info: ForkInfo,
     pub(crate) gas_used: u128,
 }
 
@@ -237,7 +226,7 @@ pub fn run_test_case(
 
     let mut state_reader = ExtendedStateReader {
         dict_state_reader: cheatnet_constants::build_testing_state(),
-        fork_state_reader: get_fork_state_reader(&runner_config.workspace_root, &case.fork_config)?,
+        fork_state_reader: get_fork_state_reader(&runner_config.workspace_root, &case.fork_config),
     };
     let block_info = state_reader.get_block_info()?;
 
@@ -289,16 +278,6 @@ pub fn run_test_case(
         extended_runtime: call_to_blockifier_runtime,
     };
 
-    let latest_block_number = if let Some(ValidatedForkConfig {
-        url: _,
-        block_id: BlockId::Tag(Latest),
-    }) = &case.fork_config
-    {
-        Some(block_info.block_number)
-    } else {
-        None
-    };
-
     let mut vm = VirtualMachine::new(true);
     let run_result = runner.run_function_with_vm(
         func,
@@ -316,9 +295,6 @@ pub fn run_test_case(
 
     Ok(RunResultWithInfo {
         run_result,
-        fork_info: ForkInfo {
-            latest_block_number,
-        },
         gas_used: gas,
     })
 }
@@ -335,7 +311,6 @@ fn extract_test_case_summary(
                     run_result,
                     case,
                     args,
-                    &result_with_info.fork_info,
                     result_with_info.gas_used,
                 )),
                 // CairoRunError comes from VirtualMachineError which may come from HintException that originates in TestExecutionSyscallHandler
@@ -347,7 +322,6 @@ fn extract_test_case_summary(
                     )),
                     arguments: args,
                     test_statistics: (),
-                    latest_block_number: result_with_info.fork_info.latest_block_number,
                 }),
                 Err(err) => bail!(err),
             }
@@ -359,7 +333,6 @@ fn extract_test_case_summary(
             msg: Some(error.to_string()),
             arguments: args,
             test_statistics: (),
-            latest_block_number: None,
         }),
     }
 }
@@ -367,30 +340,16 @@ fn extract_test_case_summary(
 fn get_fork_state_reader(
     workspace_root: &Utf8Path,
     fork_config: &Option<ValidatedForkConfig>,
-) -> Result<Option<ForkStateReader>> {
-    match fork_config {
-        Some(ValidatedForkConfig { url, mut block_id }) => {
-            if let BlockId::Tag(Latest) = block_id {
-                block_id = get_latest_block_number(url)?;
-            }
-            Ok(Some(ForkStateReader::new(
+) -> Option<ForkStateReader> {
+    fork_config
+        .as_ref()
+        .map(|ValidatedForkConfig { url, block_number }| {
+            ForkStateReader::new(
                 url.clone(),
-                block_id,
+                BlockId::Number(block_number.0),
                 Some(workspace_root.join(CACHE_DIR).as_ref()),
-            )))
-        }
-        None => Ok(None),
-    }
-}
-
-fn get_latest_block_number(url: &Url) -> Result<BlockId> {
-    let client = JsonRpcClient::new(HttpTransport::new(url.clone()));
-    let runtime = Runtime::new().expect("Could not instantiate Runtime");
-
-    match runtime.block_on(client.get_block_with_tx_hashes(BlockId::Tag(Latest))) {
-        Ok(MaybePendingBlockWithTxHashes::Block(block)) => Ok(BlockId::Number(block.block_number)),
-        _ => Err(anyhow!("Could not get the latest block number".to_string())),
-    }
+            )
+        })
 }
 
 fn get_context<'a>(runtime: &'a ForgeRuntime) -> &'a EntryPointExecutionContext {